[#3635] renamed resource type class, changed mexRetries to uint8_t

+++ nes-core/include/WorkQueues/RequestTypes/Experimental/FailQueryRequest.hpp
@@ -34,7 +34,7 @@ class FailQueryRequest : public AbstractRequest {
      */
     FailQueryRequest(NES::QueryId queryId,
                                    NES::QuerySubPlanId failedSubPlanId,
-                                   int16_t maxRetries,
+                                   uint8_t maxRetries,
                                    NES::WorkerRPCClientPtr workerRpcClient);
 
   protected:
